+++ dsp-stuff-win32/src/shared/audio_buffer.hpp
@@ -2,4 +2,6 @@
 
 #include <vector>
 
-using AudioBuffer = std::vector<double>;
+// using AudioBuffer = std::vector<double>;
+
+using AudioBuffer = std::vector<float>;
